requirements.txt
absl-py==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
chex==0.1.85 ; python_version >= "3.9" and python_version < "4.0"
cloudpickle==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.9" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.9" and python_version < "4.0"
dm-tree==0.1.8 ; python_version >= "3.9" and python_version < "4.0"
etils[epath,epy]==1.5.2 ; python_version >= "3.9" and python_version < "4.0"
flax==0.7.5 ; python_version >= "3.9" and python_version < "4.0"
fonttools==4.46.0 ; python_version >= "3.9" and python_version < "4.0"
fsspec==2023.12.1 ; python_version >= "3.9" and python_version < "4.0"
gast==0.5.4 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==7.0.0 ; python_version >= "3.9" and python_version < "3.10"
importlib-resources==6.1.1 ; python_version >= "3.9" and python_version < "4.0"
jax==0.4.21 ; python_version >= "3.9" and python_version < "4.0"
jaxlib==0.4.21 ; python_version >= "3.9" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.9" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
matplotlib==3.8.2 ; python_version >= "3.9" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.9" and python_version < "4.0"
ml-dtypes==0.3.1 ; python_version >= "3.9" and python_version < "4.0"
msgpack==1.0.7 ; python_version >= "3.9" and python_version < "4.0"
nest-asyncio==1.5.8 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.26.2 ; python_version >= "3.9" and python_version < "4.0"
opt-einsum==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
optax==0.1.7 ; python_version >= "3.9" and python_version < "4.0"
orbax-checkpoint==0.4.7 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.1.3 ; python_version >= "3.9" and python_version < "4.0"
pillow==10.1.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.25.1 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pygments==2.17.2 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rich==13.7.0 ; python_version >= "3.9" and python_version < "4.0"
scipy==1.11.4 ; python_version >= "3.9" and python_version < "4.0"
seaborn==0.13.0 ; python_version >= "3.9" and python_version < "4.0"
setuptools==69.0.2 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
soundfile==0.12.1 ; python_version >= "3.9" and python_version < "4.0"
tensorflow-probability==0.23.0 ; python_version >= "3.9" and python_version < "4.0"
tensorstore==0.1.51 ; python_version >= "3.9" and python_version < "4.0"
toolz==0.12.0 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
wfdb==4.1.2 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.9" and python_version < "4.0"
